backend/pyproject.toml
[tool.poetry]
name = "rudrax-cybersec"
version = "2.0.0"
description = "RudraX CyberSec - AI-Powered Cybersecurity Agent Stack"
authors = ["Lalit Pandit <lalittheonly@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.129.0"}
psycopg = {extras = ["binary"], version = "^3.3.2"}
pyjwt = "^2.11.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.22"
aiofiles = "^25.1.0"
httpx = "^0.28.1"
aiosqlite = "^0.22.1"
python-dotenv = "^1.2.1"
websockets = "^16.0"
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
